Add master list modal to request detailview
