+++ packages/system/src/MasterList/ListView/AppBarButtons.tsx
@@ -1,5 +1,4 @@
 import React, { FC } from 'react';
-import { MasterListSearchModal } from '@openmsupply-client/system';
 import {
   DownloadIcon,
   PrinterIcon,
@@ -8,13 +7,11 @@ import {
   ButtonWithIcon,
   Grid,
   useTranslation,
-  useToggle,
 } from '@openmsupply-client/common';
 
 export const AppBarButtons: FC = () => {
   const { info, success } = useNotification();
   const t = useTranslation('inventory');
-  const c = useToggle();
 
   return (
     <AppBarButtonsPortal>
@@ -28,16 +25,6 @@ export const AppBarButtons: FC = () => {
           Icon={<PrinterIcon />}
           label={t('button.print')}
           onClick={info('No printer detected')}
-        />
-        <ButtonWithIcon
-          Icon={<PrinterIcon />}
-          label={'Test master list modal'}
-          onClick={() => c.toggleOn()}
-        />
-        <MasterListSearchModal
-          open={c.isOn}
-          onClose={c.toggleOff}
-          onChange={n => console.log(n)}
         />
       </Grid>
     </AppBarButtonsPortal>
